requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
jinja2
aiohttp
python-multipart
requests
pydantic[email]